requirements.txt
discord.py
mcstatus
dotenv
psutil
